+++ src/Configuration/Option/Connections.php
@@ -15,10 +15,15 @@ use Predis\Configuration\OptionInterface;
 use Predis\Configuration\OptionsInterface;
 use Predis\Connection\Factory;
 use Predis\Connection\FactoryInterface;
+use Predis\Connection\PhpiredisStreamConnection;
+use Predis\Connection\PhpiredisSocketConnection;
 
 /**
- * Configures a connection factory used by the client to create new connection
- * instances for single Redis nodes.
+ * Configures a new connection factory instance.
+ *
+ * The client uses the connection factory to create the underlying connections
+ * to single redis nodes in a single-server configuration or in replication and
+ * cluster configurations.
  *
  * @author Daniele Alessandri <suppakilla@gmail.com>
  */
@@ -36,18 +41,91 @@ class Connections implements OptionInterface
         if ($value instanceof FactoryInterface) {
             return $value;
         } elseif (is_array($value)) {
-            $factory = $this->getDefault($options);
-
-            foreach ($value as $scheme => $initializer) {
-                $factory->define($scheme, $initializer);
-            }
-
-            return $factory;
+            return $this->createFactoryByArray($options, $value);
+        } elseif (is_string($value)) {
+            return $this->createFactoryByString($options, $value);
         } else {
-            $class = get_called_class();
-
-            throw new \InvalidArgumentException("$class expects a valid connection factory");
+            throw new \InvalidArgumentException(sprintf(
+                '%s expects a valid connection factory', static::class
+            ));
         }
+    }
+
+    /**
+     * Creates a new connection factory from a named array.
+     *
+     * The factory instance is configured according to the supplied named array
+     * mapping URI schemes (passed as keys) to the FCQN of classes implementing
+     * Predis\Connection\NodeConnectionInterface, or callable objects acting as
+     * lazy initalizers and returning new instances of classes implementing
+     * Predis\Connection\NodeConnectionInterface.
+     *
+     * @param OptionsInterface $options Client options
+     * @param array            $value   Named array mapping URI schemes to classes or callables
+     *
+     * @return FactoryInterface
+     */
+    protected function createFactoryByArray(OptionsInterface $options, array $value)
+    {
+        /**
+         * @var FactoryInterface
+         */
+        $factory = $this->getDefault($options);
+
+        foreach ($value as $scheme => $initializer) {
+            $factory->define($scheme, $initializer);
+        }
+
+        return $factory;
+    }
+
+    /**
+     * Creates a new connection factory from a descriptive string.
+     *
+     * The factory instance is configured according to the supplied descriptive
+     * string that identifies specific configurations of schemes and connection
+     * classes. Supported configuration values are:
+     *
+     * - "phpiredis-stream" maps tcp, redis, unix to PhpiredisStreamConnection
+     * - "phpiredis-socket" maps tcp, redis, unix to PhpiredisSocketConnection
+     * - "phpiredis" is an alias of "phpiredis-stream"
+     *
+     * @param OptionsInterface $options Client options
+     * @param string           $value   Descriptive string identifying the desired configuration
+     *
+     * @return FactoryInterface
+     */
+    protected function createFactoryByString(OptionsInterface $options, string $value)
+    {
+        /**
+         * @var FactoryInterface
+         */
+        $factory = $this->getDefault($options);
+
+        switch(strtolower($value)) {
+            case 'phpiredis':
+            case 'phpiredis-stream':
+                $factory->define('tcp', PhpiredisStreamConnection::class);
+                $factory->define('redis', PhpiredisStreamConnection::class);
+                $factory->define('unix', PhpiredisStreamConnection::class);
+                break;
+
+            case 'phpiredis-socket':
+                $factory->define('tcp', PhpiredisSocketConnection::class);
+                $factory->define('redis', PhpiredisSocketConnection::class);
+                $factory->define('unix', PhpiredisSocketConnection::class);
+                break;
+
+            case 'default':
+                return $factory;
+
+            default:
+                throw new \InvalidArgumentException(sprintf(
+                    '%s does not recognize `%s` as a supported configuration string', static::class, $value
+                ));
+        }
+
+        return $factory;
     }
 
     /**